Drop libdcadec
Is not needed anymore, and will not ever be needed again.

+++ packages/plex/mediacenter/plexmediaplayer/package.mk
@@ -67,6 +67,7 @@ if [ ! -z "$CI_CRASHDUMP_SECRET" ]; then
 fi
 
 unpack() {
+if false ; then
         if [ -d $BUILD/${PKG_NAME}-${PKG_VERSION} ]; then
           cd $BUILD/${PKG_NAME}-${PKG_VERSION} ; rm -rf build
           git pull ; git reset --hard
@@ -83,6 +84,10 @@ unpack() {
             fi
           fi
         fi
+fi
+
+           git clone /mnt/ssd/money/konvergo $BUILD/${PKG_NAME}-${PKG_VERSION}
+          cd $ROOT
 
 	if [ "$PLEX_DEBUG" = yes ]; then
 		cd $BUILD/${PKG_NAME}-${PKG_VERSION}
